requirements.txt
./assets/capepy/capepy-1.0.1-py3-none-any.whl

boto3>=1.0.0,<2.0.0
boto3-stubs>=1.0.0,<2.0.0
pulumi>=3.0.0,<4.0.0
pulumi-aws>=6.35.0,<7.0.0
pytest>=8.0.0,<9.0.0
jinja2>=3.0.0,<4.0.0
pulumi-synced-folder>=0.11.0,<1.0.0

# glue deps
aws-glue-libs @ git+https://github.com/awslabs/aws-glue-libs@9d8293962e6ffc607e5dc328e246f40b24010fa8
pandas>=2.2.0,<3.0.0
pyspark>=3.5.0,<4.0.0
# gphl etl requirements
python-dateutil>=2.9.0,<3.0.0
python-docx>=1.1.0,<2.0.0
# fast[a|q] requirements
pyfastx>=2.1.0,<3.0.0
# gphl sequencing requirements
tabula-py>=2.9.0,<3.0.0
pypdf>=4.3.0,<5.0.0
